# doc/agent-api.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Work Schedule Agent API",
    "version": "1.0.0",
    "description": "通过现有账号密码换取一小时 Agent Token，查询、创建和修改学习任务。优先级 P1 最高，截止日期精确到天。所有修改仅用于规划，不开放任务计时、完成、删除和系统管理。"
  },
  "servers": [
    {
      "url": "/api/agent/v1"
    }
  ],
  "security": [
    {
      "AgentBearer": []
    }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "loginAgent",
        "summary": "使用现有账号密码登录",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/LoginResult"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "等待秒数。"
              }
            }
          }
        }
      }
    },
    "/context": {
      "get": {
        "operationId": "getPlanningContext",
        "summary": "读取规划上下文",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Context"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "operationId": "listAgentTasks",
        "summary": "查询任务，包括已完成任务；无分页",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "todo",
                "in_progress",
                "paused",
                "done"
              ]
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "manualPriority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[1-5]$"
            }
          },
          {
            "name": "deadlineFrom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "deadlineTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Task"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createAgentTask",
        "summary": "创建任务，可同时创建最多 100 个检查点",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9._:-]{1,128}$"
            },
            "description": "推荐每次新建生成 UUID，重试使用相同键与相同请求体。24 小时内且资源仍存在时防重复，按账号、方法和请求路径区分。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "幂等重试，返回当前资源。",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "带双引号的资源 revision，可用于 If-Match。",
                "schema": {
                  "type": "string"
                }
              },
              "Idempotency-Replayed": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "true",
                    "false"
                  ]
                }
              }
            }
          },
          "201": {
            "description": "已创建",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "带双引号的资源 revision，可用于 If-Match。",
                "schema": {
                  "type": "string"
                }
              },
              "Idempotency-Replayed": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "true",
                    "false"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "operationId": "getAgentTask",
        "summary": "读取任务及检查点",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "资源 ID，作为不透明字符串处理。"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "带双引号的资源 revision，可用于 If-Match。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateAgentTask",
        "summary": "修改未完成任务的规划字段",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "资源 ID，作为不透明字符串处理。"
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\"[a-f0-9]{64}\"$"
            },
            "description": "先读取资源，将 revision 用双引号包裹，如 \"abc…\"；也可直接使用 GET 响应的 ETag。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "带双引号的资源 revision，可用于 If-Match。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "428": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}/checkpoints": {
      "post": {
        "operationId": "createAgentCheckpoint",
        "summary": "为未完成任务添加检查点",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "资源 ID，作为不透明字符串处理。"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9._:-]{1,128}$"
            },
            "description": "推荐每次新建生成 UUID，重试使用相同键与相同请求体。24 小时内且资源仍存在时防重复，按账号、方法和请求路径区分。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckpointCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "幂等重试，返回当前资源。",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Checkpoint"
                    }
                  }
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "带双引号的资源 revision，可用于 If-Match。",
                "schema": {
                  "type": "string"
                }
              },
              "Idempotency-Replayed": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "true",
                    "false"
                  ]
                }
              }
            }
          },
          "201": {
            "description": "已创建",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Checkpoint"
                    }
                  }
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "带双引号的资源 revision，可用于 If-Match。",
                "schema": {
                  "type": "string"
                }
              },
              "Idempotency-Replayed": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "true",
                    "false"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/checkpoints/{id}": {
      "get": {
        "operationId": "getAgentCheckpoint",
        "summary": "读取检查点",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "资源 ID，作为不透明字符串处理。"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Checkpoint"
                    }
                  }
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "带双引号的资源 revision，可用于 If-Match。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateAgentCheckpoint",
        "summary": "修改检查点标题、顺序及估时",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "资源 ID，作为不透明字符串处理。"
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\"[a-f0-9]{64}\"$"
            },
            "description": "先读取资源，将 revision 用双引号包裹，如 \"abc…\"；也可直接使用 GET 响应的 ETag。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckpointPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Checkpoint"
                    }
                  }
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "带双引号的资源 revision，可用于 If-Match。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "428": {
            "description": "结构化错误，包含 error.code、message、details。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AgentBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Agent 登录返回的专用 Token；不接受网页 Cookie。"
      }
    },
    "schemas": {
      "Login": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1024,
            "format": "password"
          }
        }
      },
...
      "TaskCreate": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "pattern": "\\S"
          },
          "categoryId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "tags": {
            "type": "array",
            "maxItems": 20,
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            }
          },
          "manualPriority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "description": "P1 最高，P5 最低；新建时默认 P3。"
          },
          "estimatedMinutes": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000,
            "nullable": true
          },
          "deadline": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "YYYY-MM-DD；该日结束时截止，使用服务器本地时区。null 清除截止日期。"
          },
          "checkpoints": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/CheckpointCreate"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "nullable": true,
                "additionalProperties": true
              }
            }
          }
        }
      },
      "LoginResult": {
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "expires_in"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expires_in": {
            "type": "integer",
            "enum": [
              3600
            ]
          }
        }
      },
...
      "Category": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isAnomalyBucket": {
            "type": "boolean"
          },
          "isArchiveBucket": {
            "type": "boolean"
          }
        }
      },
      "Context": {
        "type": "object",
        "required": [
          "serverTime",
          "timeZone",
          "defaultCategoryId",
          "tasks",
          "categories"
        ],
        "properties": {
          "serverTime": {
            "type": "string",
            "format": "date-time"
          },
          "timeZone": {
            "type": "string"
          },
          "defaultCategoryId": {
            "type": "string"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            },
            "description": "全部未完成任务，含检查点及 revision。"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          }
        }
      }
    }
  }
}